% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{add_workflow}
\alias{add_workflow}
\title{Add a new workflow into Senaps}
\usage{
add_workflow(name, description, organisation, groups, graph)
}
\arguments{
\item{name}{Name}

\item{description}{Description}

\item{organisation}{Organisation}

\item{groups}{Groups}

\item{graph}{Graph}
}
\description{
Add a new workflow into Senaps
}
