% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{clean_group}
\alias{clean_group}
\title{Delete a group and all children with all streams, locations and platforms associated. Use it with your own risk.}
\usage{
clean_group(id, ask = TRUE)
}
\arguments{
\item{id}{Group id}

\item{ask}{Logical. Whether to ask for confirmation or not.}
}
\description{
Delete a group and all children with all streams, locations and platforms associated. Use it with your own risk.
}
