% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platforms.R
\name{current_deployment}
\alias{current_deployment}
\title{Get the current deployment for a platform}
\usage{
current_deployment(platform_detail)
}
\arguments{
\item{platform_detail}{A list of platform details. Output from `get_platform`}
}
\value{
A list with deployment details
}
\description{
Get the current deployment for a platform
}
\examples{
\dontrun{
  current_deployment('rsenaps.platform.test1')
}

}
