% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{delete_group}
\alias{delete_group}
\title{Delete a group}
\usage{
delete_group(id, cascade = FALSE)
}
\arguments{
\item{id}{Group ID}

\item{cascade}{Remove group even when not empty (default FALSE)}
}
\value{
The staus code of delete
}
\description{
Delete a group
}
