% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{delete_location}
\alias{delete_location}
\title{Delete an existing Location.}
\usage{
delete_location(id, cascade = FALSE)
}
\arguments{
\item{id}{The location id}

\item{cascade}{Remove Location even when is referenced by other objects (eg. Platform deployments or Streams).}
}
\description{
Delete an existing Location.
}
