% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{delete_observations}
\alias{delete_observations}
\title{Delete all observations in a stream.}
\usage{
delete_observations(id)
}
\arguments{
\item{id}{The stream id}
}
\value{
The status code of request
}
\description{
Delete all observations in a stream.
}
