% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platforms.R
\name{delete_platform}
\alias{delete_platform}
\title{Delete an existing platform}
\usage{
delete_platform(id, cascade = FALSE)
}
\arguments{
\item{id}{The platfrom id}

\item{cascade}{Logical. Should the deletion take place even if the platform is referenced by other objects.}
}
\value{
The status code of request.
\itemize{
  \item 200 OK
  \item 400 The Platform is referenced by other objects. Either use the 'cascade' parameter or manually remove references.
  \item 401 Unauthorised to delete this Platform
  \item 404 Not found
}
}
\description{
Delete an existing platform
}
\examples{
\dontrun{
  delete_platform('rsenaps.platform.test1')
}

}
