% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedules.R
\name{delete_schedule}
\alias{delete_schedule}
\title{Delete an existing schedule}
\usage{
delete_schedule(id)
}
\arguments{
\item{id}{The schedule id}
}
\value{
The status code of request.
\itemize{
  \item 200 OK
  \item 401 Unauthorised to delete this Platform
  \item 404 Not found
}
}
\description{
Delete an existing schedule
}
