% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streams.R
\name{delete_stream}
\alias{delete_stream}
\title{Delete an existing stream}
\usage{
delete_stream(id)
}
\arguments{
\item{id}{The stream id}
}
\value{
The status code of request
}
\description{
Delete an existing stream
}
