% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{delete_workflow}
\alias{delete_workflow}
\title{Delete an existing workflow}
\usage{
delete_workflow(id)
}
\arguments{
\item{id}{The platfrom id}
}
\value{
The status code of request.
\itemize{
  \item 200 OK
  \item 401 Unauthorised to delete this Platform
  \item 404 Not found
}
}
\description{
Delete an existing workflow
}
