% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senaps.R
\name{format_datetime}
\alias{format_datetime}
\title{Format data and time for Senaps}
\usage{
format_datetime(dt, tz = "GMT", url_encode = FALSE)
}
\arguments{
\item{dt}{Date and time format}

\item{tz}{timezone}

\item{url_encode}{Whether to use RUL encode}
}
\value{
A string for Senaps
}
\description{
Format data and time for Senaps
}
