% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{get_aggregation}
\alias{get_aggregation}
\title{Calculate an aggregated view of observations. The aggregation is calculated on demand.}
\usage{
get_aggregation(
  streamid,
  aggperiod,
  start = NULL,
  end = NULL,
  si = TRUE,
  ei = TRUE,
  tz = "GMT",
  limit = 1000,
  name_streams = FALSE
)
}
\arguments{
\item{streamid}{Stream identifier or a comma separated list of stream identifiers}

\item{aggperiod}{The number of milliseconds in each aggregation interval}

\item{start}{Start date}

\item{end}{End date}

\item{si}{Is the start parameter treated as an inclusive boundary}

\item{ei}{Is the end parameter treated as an inclusive boundary}

\item{tz}{The timezone of timestamp. GMT/UTC time in default}

\item{limit}{Limit the number of results. The limit is 1000 by default.}
}
\description{
Calculate an aggregated view of observations. The aggregation is calculated on demand.
}
