% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseimages.R
\name{get_baseimage}
\alias{get_baseimage}
\title{The base image in the Senaps}
\usage{
get_baseimage(id)
}
\arguments{
\item{id}{A baseimage id}
}
\value{
A list for base image
}
\description{
The base image in the Senaps
}
