% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseimages.R
\name{get_baseimages}
\alias{get_baseimages}
\title{The all base images in the Senaps}
\usage{
get_baseimages()
}
\value{
A list of base images
}
\description{
The all base images in the Senaps
}
