% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_data}
\alias{get_data}
\title{Retrieving data from the NetCDF Subset Service (NCSS) in Senaps}
\usage{
get_data(path, var, latitude = NULL, longitude = NULL,
  temporal = "all", format = NULL, tz = "UTC", time_var = NULL)
}
\arguments{
\item{path}{The URL path at which to place the data.}

\item{var}{Name of variables.}

\item{latitude}{Point location. units of degrees_east, degrees_north}

\item{longitude}{Point location. units of degrees_east, degrees_north}

\item{temporal}{all}

\item{time_var}{The variable name of time.}

\item{accept}{Used to specify the returned format.
Grid: netCDF and netCDF4.
Point: netCDF, netCDF4, XML, CSV.
Station: netCDF, netCDF4, XML, CSV, WaterML2.}
}
\value{
The data in the NCSS
}
\description{
See details from https://www.unidata.ucar.edu/software/thredds/current/tds/reference/NetcdfSubsetServiceReference.html
}
