% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{get_group}
\alias{get_group}
\title{The group detail information in the Senaps}
\usage{
get_group(id, recursive = FALSE)
}
\arguments{
\item{id}{The group id}

\item{recursive}{Return full details of embedded resources}
}
\value{
list of group meta data. NULL if group doesn't exist
}
\description{
The group detail information in the Senaps
}
