% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{get_groups}
\alias{get_groups}
\title{The all groups in the Senaps}
\usage{
get_groups(id = NULL, organisation = NULL, groups = NULL,
  expand = FALSE, usermetadatafield = NULL,
  usermetadatavalues = NULL)
}
\value{
A data frame with all groups
}
\description{
The all groups in the Senaps
}
