% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{get_job_log}
\alias{get_job_log}
\title{Get logs for a jog}
\usage{
get_job_log(jobid, limit = NULL)
}
\arguments{
\item{jobid}{The job id}

\item{limit}{Number of logs retrieved}
}
\value{
A list for job
}
\description{
Get logs for a jog
}
