% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{get_jobs}
\alias{get_jobs}
\title{The all jobs for a workflow in the Senaps}
\usage{
get_jobs(workflowid = NULL, limit = NULL, skip = NULL)
}
\arguments{
\item{workflowid}{Workflow id}

\item{limit}{Number of jobs retrieved}

\item{skip}{Skip the number of jobs for pagination}
}
\value{
A list of jobs
}
\description{
The all jobs for a workflow in the Senaps
}
