% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{get_location}
\alias{get_location}
\title{The detail inforamtion of locations in the Senaps}
\usage{
get_location(id)
}
\arguments{
\item{id}{Whether return detail information of locations}
}
\value{
A list of location meta information
}
\description{
The detail inforamtion of locations in the Senaps
}
