% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{get_locations}
\alias{get_locations}
\title{The all locations in the Senaps}
\usage{
get_locations(groups = NULL)
}
\value{
A data.frame of locations
}
\description{
The all locations in the Senaps
}
