% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{get_me}
\alias{get_me}
\title{Get the authenticated user using root api call
:reference: https://data.sense-t.org.au/api/sensor/v2/api-docs/#!/default/}
\usage{
get_me()
}
\description{
Get the authenticated user using root api call
:reference: https://data.sense-t.org.au/api/sensor/v2/api-docs/#!/default/
}
