% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_model}
\alias{get_model}
\title{The model detail information in the Senaps}
\usage{
get_model(id)
}
\arguments{
\item{id}{A model id}
}
\value{
A list of model meta data
}
\description{
The model detail information in the Senaps
}
