% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{get_observations}
\alias{get_observations}
\title{Get a collection of observations}
\usage{
get_observations(
  streamid,
  start = NULL,
  end = NULL,
  si = TRUE,
  ei = TRUE,
  tz = "GMT",
  descending = FALSE,
  limit = 1000,
  name_streams = FALSE
)
}
\arguments{
\item{streamid}{Stream identifier or a comma separated list of stream identifiers}

\item{start}{Start date}

\item{end}{End date}

\item{si}{Is the start parameter treated as an inclusive boundary}

\item{ei}{Is the end parameter treated as an inclusive boundary}

\item{tz}{The timezone of timestamp. GMT/UTC time in default}

\item{descending}{Sort the results. By default results are returned in ascending order.}

\item{limit}{Limit the number of results. The limit is 1000 by default.}

\item{name_streams}{when TRUE, the returned value column name is the stream id which will facilitate future queries to return multiple streams.}
}
\description{
Get a collection of observations
}
