% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platforms.R
\name{get_platform}
\alias{get_platform}
\title{Get details about a platform.}
\usage{
get_platform(id)
}
\arguments{
\item{id}{The platform id}
}
\value{
A list of platform details. NULL if platform does not exist.
}
\description{
Get details about a platform.
}
\examples{
\dontrun{
  get_platform('rsenaps.platform.test1')
}

}
