% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platforms.R
\name{get_platforms}
\alias{get_platforms}
\title{The all platforms in the Senaps}
\usage{
get_platforms(groups = NULL)
}
\arguments{
\item{groups}{A list of group ids}
}
\value{
A data frame with all streamids
}
\description{
The all platforms in the Senaps
}
\examples{
\dontrun{
  get_platforms()
}

}
