% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roles.R
\name{get_role}
\alias{get_role}
\title{The detail inforamtion of role in the Senaps}
\usage{
get_role(id)
}
\arguments{
\item{id}{Role id}
}
\value{
A list of role meta information
}
\description{
The detail inforamtion of role in the Senaps
}
