% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roles.R
\name{get_roles}
\alias{get_roles}
\title{Get roles in Senaps}
\usage{
get_roles(groups = NULL)
}
\arguments{
\item{groups}{The group ids}
}
\value{
A character vector of role ids
}
\description{
Get roles in Senaps
}
