% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedules.R
\name{get_schedule}
\alias{get_schedule}
\title{The detail inforamtion of schedule in the Senaps}
\usage{
get_schedule(id)
}
\arguments{
\item{id}{schedule id}
}
\value{
A list of schedule meta information
}
\description{
The detail inforamtion of schedule in the Senaps
}
