% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedules.R
\name{get_schedules}
\alias{get_schedules}
\title{Get schedules in Senaps}
\usage{
get_schedules(workflow = NULL)
}
\arguments{
\item{workflow}{The workflow id}
}
\value{
A character vector of role ids
}
\description{
Get schedules in Senaps
}
