% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streams.R
\name{get_stream}
\alias{get_stream}
\title{The stream detail information in the Senaps}
\usage{
get_stream(id)
}
\arguments{
\item{id}{The stream id}
}
\value{
A list of steam meta information
}
\description{
The stream detail information in the Senaps
}
