% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{get_workflow}
\alias{get_workflow}
\title{The workflow detail information in the Senaps}
\usage{
get_workflow(id)
}
\value{
A list of workflow meta data
}
\description{
The workflow detail information in the Senaps
}
