% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{get_workflows}
\alias{get_workflows}
\title{The all workflows in the Senaps}
\usage{
get_workflows(name = NULL, groups = NULL)
}
\value{
A list of workflows
}
\description{
The all workflows in the Senaps
}
