% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-execution.R
\name{local_senaps}
\alias{local_senaps}
\title{Test if execution is local}
\usage{
local_senaps()
}
\value{
logical
}
\description{
This function looks for the environment variable SENAPS_LOCAL, which must be set for the function to return TRUE
}
\details{
The easiest way to set the variable is with the helper function \code{set_local()}, which will set
write the variable into a .Renviron file.

Alternatively, the variable can be set in a session with \code{Sys.setenv("SENAPS_LOCAL"=TRUE)},
or manually added to a .Renviron file.
}
\examples{
\donttest{
  if(local_senaps()) {
    # some logic to do locally
  }
}
}
