% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senaps.R
\name{parse_parameter}
\alias{parse_parameter}
\title{Parse the parameter in a string with key and value}
\usage{
parse_parameter(p)
}
\arguments{
\item{p}{The parameter}
}
\value{
A list of parameter values
}
\description{
Parse the parameter in a string with key and value
}
