% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{put_group}
\alias{put_group}
\title{Create a new group into Senaps}
\usage{
put_group(id, name, description, organisation, groups = NULL,
  usermetadata = NULL)
}
\arguments{
\item{id}{The new group unique id}

\item{name}{The name of group}

\item{description}{The description of group}

\item{organisation}{The organisation ids}

\item{groups}{The parent groups}

\item{usermetadata}{A list of usermeta information}
}
\value{
TRUE if new group is added
}
\description{
Create a new group into Senaps
}
