% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{put_location}
\alias{put_location}
\title{Create a new location into Senaps}
\usage{
put_location(id, description, organisation, longitude, latitude,
  elevation = NULL, groups = character(), usermetadata = NULL)
}
\arguments{
\item{id}{The id of new location. Have to be unique}

\item{description}{The description of new location}

\item{organisation}{The organisation of new location. Have to be an existing organisation}

\item{longitude}{The longitude (degree)}

\item{latitude}{The latitude (degree)}

\item{elevation}{The elevation above sea level (m, optional)}

\item{groups}{The list of groups. Have to be an existing group (optional)}
}
\value{
A list of new location if the new location is successfully created
}
\description{
Create a new location into Senaps
}
