% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{put_model}
\alias{put_model}
\title{Upload a model into Senaps}
\usage{
put_model(archive)
}
\arguments{
\item{archive}{The file path to a model archive with file type zip or tar.gz}
}
\value{
TRUE if model is uploaded
}
\description{
Upload a model into Senaps
}
