% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{put_observations}
\alias{put_observations}
\title{Add new observations into a stream}
\usage{
put_observations(id, timestamp, value)
}
\arguments{
\item{id}{The stream id}

\item{timestamp}{The timestamp vector of observations}

\item{value}{The observed values}
}
\description{
Add new observations into a stream
}
