% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platforms.R
\name{put_platform}
\alias{put_platform}
\title{Create a new platform or update an existing platform}
\usage{
put_platform(
  id,
  name = id,
  organisation,
  groups = NULL,
  streams = NULL,
  deployments = NULL,
  usermetadata = NULL
)
}
\arguments{
\item{id}{Platform ID. Have to be unique}

\item{name}{Platform name}

\item{organisation}{Platform organisation. Have to be an existing organisation}

\item{groups}{Platform groups. Have to be an existing groups (optional)}

\item{streams}{Platform streams Have to be an existing streams (optional)}

\item{deployments}{Platform deployments (optional)}

\item{usermetadata}{Arbitrary usermetadata. Any valid list.}
}
\value{
TRUE if sucessful.
}
\description{
Create a new platform or update an existing platform
}
\examples{
\dontrun{
 id <- 'rsenaps.platform.test1'
 name <- "RSenaps Platform Test1"
 organisation <- "sandbox"
 put_platform(id =  id, name = name, organisation = organisation)
}

}
