% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roles.R
\name{put_role}
\alias{put_role}
\title{Add a new role}
\usage{
put_role(id, type, organisation, group, permissions)
}
\arguments{
\item{id}{Role id}

\item{type}{Role type}

\item{organisation}{Organisation}

\item{group}{Group}

\item{permissions}{A vector of of permissions}
}
\description{
Add a new role
}
