% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedules.R
\name{put_schedule}
\alias{put_schedule}
\title{Add a new schedule}
\usage{
put_schedule(
  name,
  workflow,
  cron = "0 */1 * * *",
  timezone = "UTC",
  description = name,
  active = TRUE
)
}
\arguments{
\item{name}{The schedule name}

\item{workflow}{A workflow id}

\item{cron}{A cron expression}

\item{timezone}{Timezone}

\item{description}{Description of the schedule}

\item{active}{Whether the schedule is active or not}
}
\value{
The schedule id
}
\description{
Add a new schedule
}
