% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streams.R
\name{put_stream}
\alias{put_stream}
\title{Create a new stream in Senaps}
\usage{
put_stream(
  id,
  organisation,
  reportingPeriod,
  samplePeriod,
  observedProperty,
  unitOfMeasure,
  location = NULL,
  groups = character(),
  resulttype = "scalarvalue",
  interpolationType = "Continuous",
  usermetadata = NULL
)
}
\arguments{
\item{id}{The id of new stream. Have to be unique}

\item{organisation}{An organisation id}

\item{reportingPeriod}{Reporting frequency. TODO: Add documentation about valid format}

\item{samplePeriod}{Sampling frequency. TODO: Add documentation about valid format}

\item{observedProperty}{Observe property. TODO: Add documentation about valid values}

\item{unitOfMeasure}{Observe unit. TODO: Add documentation about valid values}

\item{location}{The location of new location. Have to be an existing location}

\item{groups}{The list of groups. Have to be an existing group (optional)}

\item{resulttype}{Types of reuls. scalarvalue in default}

\item{interpolationType}{Interpolation type. Continuous in default}

\item{usermetadata}{User metadata values}
}
\value{
A list of new stream if the new stream is successfully created
}
\description{
Create a new stream in Senaps
}
