% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{get_groups}
\alias{get_groups}
\title{The all groups in the Senaps}
\usage{
get_groups(
  id = NULL,
  organisation = NULL,
  groups = NULL,
  expand = FALSE,
  usermetadatafield = NULL,
  usermetadatavalues = NULL
)
}
\arguments{
\item{id}{Only return groups with this id or that match the pattern. Accepts * and ? wildcards}

\item{organisation}{An organisation id}

\item{groups}{A list of group ids to filter by}

\item{expand}{Logical. Should the response object be expanded?}

\item{usermetadatafield}{The field in the usermetadata you'd like to filter by}

\item{usermetadatavalues}{Values to filter the response by in the usermetadatafield}
}
\value{
A character vector or named list, depending on expand value
}
\description{
The all groups in the Senaps
}
