% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platforms.R
\name{get_platforms}
\alias{get_platforms}
\title{The all platforms in the Senaps}
\usage{
get_platforms(id = NULL, groupids = NULL, groups = NULL)
}
\arguments{
\item{id}{Only return platforms with this id or partial match using wildcards (*, ?).}

\item{groupids}{filter response by a comma separated list of group ids}

\item{groups}{filter response by a comma separated list of group ids}
}
\value{
A data frame with all streamids
}
\description{
The all platforms in the Senaps
}
\examples{
\dontrun{
  get_platforms()
}

}
