% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streams.R
\name{get_streams}
\alias{get_streams}
\title{The all streams in the Senaps}
\usage{
get_streams(groups = NULL, observed_property = NULL, limit = 1000)
}
\arguments{
\item{groups}{A group id (optional)}

\item{observed_property}{An observed property uri}

\item{limit}{Maximum records to retrieve}
}
\value{
A data frame with all streamids
}
\description{
The all streams in the Senaps
}