% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedules.R
\name{put_schedule}
\alias{put_schedule}
\title{Add a new schedule}
\usage{
put_schedule(name, workflow, cron = "0 */1 * * *", timezone = "UTC",
  description = name, active = TRUE)
}
\arguments{
\item{id}{schedule id}

\item{type}{schedule type}

\item{organisation}{Organisation}

\item{group}{Group}

\item{permissions}{A vector of of permissions}
}
\value{
The schedule id
}
\description{
Add a new schedule
}
