% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{senaps_options}
\alias{senaps_options}
\title{Set or get options for my package}
\usage{
senaps_options(...)
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]} pairs to set options.}
}
\description{
Set or get options for my package
}
\section{Supported options}{

The following options are supported
\itemize{
 \item{\code{sensor_url}}{ The url for sensor API}
 \item{\code{analysis_url}}{ The url for analysis API}
 \item{\code{username}}{ The username for Senaps}
 \item{\code{password}}{ The password for Senaps}
 \item{\code{apikey}}{ The apikey for Senaps}
}
}

